requirements.txt
mediapipe
opencv-python
pynput
pyyaml
